--- test/spec/geo/geometry-views/base/fake-multi-point-view.js
@@ -1,8 +1,0 @@
-var MultiGeometryViewBase = require('../../../../../src/geo/geometry-views/base/multi-geometry-view-base');
-var PointView = require('./fake-point-view');
-
-var MultiPointView = MultiGeometryViewBase.extend({
-  GeometryViewClass: PointView
-});
-
-module.exports = MultiPointView;
