--- test/spec/geo/geometry-views/base/point-view-spec.js
@@ -1,7 +1,0 @@
-var MapView = require('./fake-map-view');
-var PointView = require('./fake-point-view');
-var SharedTestsForPointViews = require('../shared-tests-for-point-views');
-
-describe('src/geo/geometry-views/base/point-view.js', function () {
-  SharedTestsForPointViews.call(this, MapView, PointView);
-});
