+++ test/spec/geo/geometry-views/gmaps/multi-point-view-spec.js
@@ -1,4 +1,4 @@
-var GMapsMapView = require('./gmaps-map-for-tests');
+var GMapsMapView = require('../../../../../src/geo/gmaps/gmaps-map-view.js');
 var GMapsMultiPointView = require('../../../../../src/geo/gmaps/geometries/multi-point-view.js');
 var SharedTestsForMultiPointViews = require('../shared-tests-for-multi-point-views');
