+++ test/spec/geo/geometry-views/gmaps/multi-polygon-view-spec.js
@@ -1,4 +1,4 @@
-var GMapsMapView = require('./gmaps-map-for-tests');
+var GMapsMapView = require('../../../../../src/geo/gmaps/gmaps-map-view.js');
 var GMapsMultiPolygonView = require('../../../../../src/geo/gmaps/geometries/multi-polygon-view.js');
 var SharedTestsForMultiPolygonViews = require('../shared-tests-for-multi-polygon-views');
