+++ test/spec/geo/geometry-views/gmaps/multi-polyline-view-spec.js
@@ -1,4 +1,4 @@
-var GMapsMapView = require('./gmaps-map-for-tests');
+var GMapsMapView = require('../../../../../src/geo/gmaps/gmaps-map-view.js');
 var GMapsMultiPolylineView = require('../../../../../src/geo/gmaps/geometries/multi-polyline-view.js');
 var SharedTestsForMultiPolylineViews = require('../shared-tests-for-multi-polyline-views');
