+++ test/spec/geo/geometry-views/gmaps/point-view-spec.js
@@ -1,4 +1,4 @@
-var GMapsMapView = require('./gmaps-map-for-tests');
+var GMapsMapView = require('../../../../../src/geo/gmaps/gmaps-map-view.js');
 var GMapsPointView = require('../../../../../src/geo/gmaps/geometries/point-view.js');
 var SharedTestsForPointViews = require('../shared-tests-for-point-views');
