+++ test/spec/geo/geometry-views/gmaps/polyline-view-spec.js
@@ -1,4 +1,4 @@
-var GMapsMapView = require('./gmaps-map-for-tests');
+var GMapsMapView = require('../../../../../src/geo/gmaps/gmaps-map-view.js');
 var GMapsPolylineView = require('../../../../../src/geo/gmaps/geometries/polyline-view.js');
 var SharedTestsForPolylineViews = require('../shared-tests-for-polyline-views');
