+++ test/spec/geo/geometry-views/leaflet/multi-point-view-spec.js
@@ -1,4 +1,4 @@
-var LeafletMapView = require('./leaflet-map-view-for-tests');
+var LeafletMapView = require('../../../../../src/geo/leaflet/leaflet-map-view.js');
 var LeafletMultiPointView = require('../../../../../src/geo/leaflet/geometries/multi-point-view.js');
 var SharedTestsForMultiPointViews = require('../shared-tests-for-multi-point-views');
