+++ test/spec/geo/geometry-views/leaflet/multi-polygon-view-spec.js
@@ -1,4 +1,4 @@
-var LeafletMapView = require('./leaflet-map-view-for-tests');
+var LeafletMapView = require('../../../../../src/geo/leaflet/leaflet-map-view.js');
 var LeafletMultiPolygonView = require('../../../../../src/geo/leaflet/geometries/multi-polygon-view.js');
 var SharedTestsForMultiPolygonViews = require('../shared-tests-for-multi-polygon-views');
