+++ test/spec/geo/geometry-views/leaflet/point-view-spec.js
@@ -1,4 +1,4 @@
-var LeafletMapView = require('./leaflet-map-view-for-tests');
+var LeafletMapView = require('../../../../../src/geo/leaflet/leaflet-map-view.js');
 var LeafletPointView = require('../../../../../src/geo/leaflet/geometries/point-view.js');
 var SharedTestsForPointViews = require('../shared-tests-for-point-views');
