+++ test/spec/geo/geometry-views/leaflet/polyline-view-spec.js
@@ -1,4 +1,4 @@
-var LeafletMapView = require('./leaflet-map-view-for-tests');
+var LeafletMapView = require('../../../../../src/geo/leaflet/leaflet-map-view.js');
 var LeafletPolylineView = require('../../../../../src/geo/leaflet/geometries/polyline-view.js');
 var SharedTestsForPolylineViews = require('../shared-tests-for-polyline-views');
